Creacion de contratacion desde cliente

+++ fixer-app-backend/src/main/java/org/example/fixerappbackend/model/Usuario.java
@@ -9,7 +9,7 @@ public abstract class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
-    private Integer id;
+    private Long id;
 
     @Column(name = "nombre")
     private String nombre;
@@ -28,8 +28,8 @@ public abstract class Usuario {
 
     // Getters y setters
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
     public String getNombreUsuario() { return nombreUsuario; }
